Created RubyBugzilla::Service class
